fix type of the variable

+++ includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: aabda <aabda@student.s19.be>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/30 14:06:36 by aabda             #+#    #+#             */
-/*   Updated: 2022/08/03 20:41:01 by aabda            ###   ########.fr       */
+/*   Updated: 2022/08/03 20:50:58 by aabda            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ int		ft_is_percent(va_list arg);
 int		ft_is_pointer(va_list arg);
 int		ft_count_int(int nb);
 int		ft_count_unsigned(unsigned int nb);
-int		ft_count_hex(int nb);
+int		ft_count_hex(unsigned long nb);
 int		ft_search_tc(const char *format, va_list arg);
 size_t	ft_strlen(const char *s);
 void	ft_putnbr_hex_fd(const char format, unsigned int n, int fd);
